Fix adding [1] clause to ID queries with monitors
We need to traverse inside the monitor to properly check if the
index is already there, and to avoid unnormalized ThingTalk with
[1] on the outside of a monitor.
